refactor(web): consolidate raw-fetch Vellum-Organization-Id + CSRF header injection (LUM-1787)
Five files under apps/web/src that use raw `fetch('/v1/...')` for the
Django wildcard proxy were each hand-rolling the `Vellum-Organization-Id`
(and `X-CSRFToken` for mutations) header block inline. Each one was
correct in isolation, but the pattern was drift-prone — the next time a
new raw-fetch call lands, the next maintainer has to remember to add the
headers, and the silent failure mode (Django returns 400 from
`MissingVellumOrganizationIDHeaderError`, the caller's null-fallback
kicks in, UI degrades quietly) is exactly what LUM-1784 looked like.

Extract `buildVellumHeaders()` and `buildVellumMutatingHeaders()` from
the local helpers in `oauth-apps.ts` into a top-level shared utility at
`apps/web/src/lib/auth/request-headers.ts`. Replace inline header
construction at every call site:

  - components/share-feedback-modal.tsx
  - domains/chat/components/chat-attachments/attachment-preview-modal.tsx
  - domains/settings/api/oauth-apps.ts        (also drops the local helpers)
  - domains/settings/api/oauth-providers.ts
  - domains/settings/components/panels/doctor-panel.tsx (two sites:
    `doctorFetch` and the SSE consumer)

None of these endpoints are in the daemon's OpenAPI schema (they're
served through Django's wildcard `RuntimeProxyView`), so the generated
HeyAPI client can't reach them — raw fetch + the shared helper is the
right shape. Anywhere new that genuinely needs raw fetch should import
from `lib/auth/request-headers.js`; anything else should keep using the
generated client where the interceptor handles this automatically.

Touches: 1 new file, 5 updated, 0 removed. Behavior unchanged — same
headers attached to same requests.
